badruimte_met_toilet verplaatst naar woningwaardering/vera/utils.py en gerefactored
